+++ app/Http/Controllers/TdpController.php
@@ -50,21 +50,32 @@ class TdpController extends Controller
 
             'MasaBerlaku' => 'required_if:MasaBerlakuStatus,==,1',
             
-            'FileTdp' => 'required|mimes:pdf|max:2000',
+            'FileTdp' => 'mimes:pdf|max:6000',
             
             ]);
     
-            // menyimpan data file yang diupload ke variabel $file
-            $file = $request->file('FileTdp');
-            $extension = $file->getClientOriginalExtension();
-            $destination_folder = public_path() . DIRECTORY_SEPARATOR . 'documents' . DIRECTORY_SEPARATOR .'tdp';
+            $tdp = new Tdp;
+
+            if($request->hasFile('FileTdp')){
+                // menyimpan data file yang diupload ke variabel $file
+                $file = $request->file('FileTdp');
+                $extension = $file->getClientOriginalExtension();
+                $destination_folder = public_path() . DIRECTORY_SEPARATOR . 'documents' . DIRECTORY_SEPARATOR .'tdp';
     
+                $fileName = md5(time()) . '.' . $extension;;
+                // $fileName = $akta->getNextId();
+                $tdp->filename = $fileName;
+            
+                // upload file
+                $file->move($destination_folder,$fileName);
+            }
+            
             $vendor_id = $request->input('vendor_id');
             $vendor_name = $request->input('vendor_name');
             $tgl_tdp = $request->input('TanggalPenerbitan');
             $tgl_berlaku = $request->input('MasaBerlaku');
     
-            $tdp = new Tdp;
+            
             $tdp->no_dokumen = $request->input('NomorDokumen');
             $tdp->vendor_id = $vendor_id;
             $tdp->tgl_penerbitan = Carbon::parse($tgl_tdp)->format('Y-m-d H:i:s');
@@ -72,12 +83,7 @@ class TdpController extends Controller
             $tdp->masa_berlaku_status = $request->input('MasaBerlakuStatus');
             $tdp->berlaku_sampai = Carbon::parse($tgl_berlaku)->format('Y-m-d H:i:s');
     
-            $fileName = md5(time()) . '.' . $extension;;
-            // $fileName = $akta->getNextId();
-            $tdp->filename = $fileName;
             
-            // upload file
-            $file->move($destination_folder,$fileName);
     
             $tdp->save();
             //return "store";
@@ -120,7 +126,7 @@ class TdpController extends Controller
             'NomorDokumen' => 'required',
             'TanggalPenerbitan' => 'required',
             'InstansiPenerbit' => 'required',
-
+            'FileTdp' => 'mimes:pdf|max:6000',
             'MasaBerlaku' => 'required_if:MasaBerlakuStatus,==,1',
             
             ]);
@@ -145,7 +151,7 @@ class TdpController extends Controller
             //Delete file yang sebelumnya
             $file_path = public_path().'/documents/tdp/'.$tdp->filename;
             $check_file = File::exists($file_path);
-            if ($check_file){
+            if ($check_file && $tdp->filename<>""){
                 unlink($file_path);
             }
 
@@ -177,7 +183,7 @@ class TdpController extends Controller
         $vendor_id = $tdp->vendor_id;
         $file_path = public_path().'/documents/tdp/'.$tdp->filename;
         $check_file = File::exists($file_path);
-        if ($check_file){
+        if ($check_file && $tdp->filename<>""){
             unlink($file_path);
         }
         //File::delete($file_path);
